requirements.txt
attrs==24.2.0
cads-api-client==1.4.5
cdsapi==0.7.3
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
eccodes==2.38.3
exceptiongroup==1.2.2
fast-histogram==0.14
findlibs==0.0.5
fonttools==4.54.1
gsw==3.6.19
idna==3.10
imageio==2.36.0
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
mpl-scatter-density==0.7
multiurl==0.3.1
netCDF4==1.7.1.post2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pycparser==2.22
pygrib==2.1.6
pyparsing==3.2.0
pyproj==3.7.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xarray==2024.9.0
